Add .bacpac download, improve backup/restore UX
- Add API endpoint to generate and stream .bacpac database backups
- Add endpoints for backup plan and Azure SQL export trigger
- Switch client downloads to JS interop for direct file saving
- Update API client to return file content, not just URLs
- Enhance UI feedback and grid display for backup/restore
- Fix SizeBytes type from Int64 to Int32 in backup queries
- Add FileDownloadDto and improve error handling/logging

+++ Budget.Api/Features/Utilities/ImportExport/GetBackupSetDetails.cs
@@ -55,7 +55,7 @@ public static class GetBackupSetDetails
       {
         var tableName = entity.RowKey;
         var blobName = entity.GetString("BlobName") ?? string.Empty;
-        var sizeBytes = entity.GetInt64("SizeBytes") ?? 0;
+        var sizeBytes = entity.GetInt32("SizeBytes") ?? 0;
         var exportedAt = entity.GetDateTime("ExportedAt") ?? DateTime.MinValue;
 
         tables.Add(new BackupTableDto(tableName, blobName, sizeBytes, exportedAt, request.PartitionKey));
